+++ src/utils/object.ts
@@ -1,7 +1,9 @@
+/* eslint-disable prefer-object-has-own */
+
 /**
  * Creates an object composed of the picked object properties, like `lodash.pick`.
  */
-export function pick<T extends Record<string, any>, K extends keyof T>(obj: T, keys: (K | string)[]): Pick<T, K>
+export function pick<T extends Pick<T, K>, K extends keyof T>(obj: T, keys: K[]): Pick<T, K>
 {
     return keys.reduce((acc, key) =>
     {
@@ -10,5 +12,5 @@ export function pick<T extends Record<string, any>, K extends keyof T>(obj: T, k
             acc[key] = obj[key];
         }
         return acc;
-    }, {} as any);
+    }, {} as Pick<T, K>);
 }